+++ sorts.h
@@ -1,8 +1,8 @@
 #ifndef SORTS_H
 #define SORTS_H
 
-void selectionSort(double* array, int len);
+void selectionSort(long* array, int len);
 
-void visualizeArray(double* array, int len);
+void visualizeArray(long* array, int len);
 
 #endif
