+++ visualizeArray.c
@@ -1,13 +1,13 @@
 #include <stdio.h>
 
-void visualizeArray(double* array, int len)
+void visualizeArray(long* array, int len)
 {
     for (int i = 0; i < len; i++)
     {
-        printf("%i ", array[i]);
+        printf("%li ", array[i]);
         for (int j = 0; j < array[i]; j++)
         {
-            printf('#');
+            printf("#");
         }
         printf("\n");
     }
